+++ dom/src/main/java/react4j/dom/events/ClipboardEventHandler.java
@@ -3,6 +3,7 @@ package react4j.dom.events;
 import jsinterop.annotations.JsFunction;
 
 @JsFunction
+@FunctionalInterface
 public interface ClipboardEventHandler
 {
   void onClipboardEvent( ClipboardEvent event );