+++ dom/src/main/java/react4j/dom/events/DragEventHandler.java
@@ -3,6 +3,7 @@ package react4j.dom.events;
 import jsinterop.annotations.JsFunction;
 
 @JsFunction
+@FunctionalInterface
 public interface DragEventHandler
 {
   public void onDragEvent( DragEvent event );