+++ dom/src/main/java/react4j/dom/events/ReactEventHandler.java
@@ -3,6 +3,7 @@ package react4j.dom.events;
 import jsinterop.annotations.JsFunction;
 
 @JsFunction
+@FunctionalInterface
 public interface ReactEventHandler
 {
   void onReactEvent( SyntheticEvent event );