+++ dom/src/main/java/react4j/dom/events/TouchEventHandler.java
@@ -3,6 +3,7 @@ package react4j.dom.events;
 import jsinterop.annotations.JsFunction;
 
 @JsFunction
+@FunctionalInterface
 public interface TouchEventHandler
 {
   void onTouchEvent( TouchEvent event );