+++ dom/src/main/java/react4j/dom/events/WheelEventHandler.java
@@ -3,6 +3,7 @@ package react4j.dom.events;
 import jsinterop.annotations.JsFunction;
 
 @JsFunction
+@FunctionalInterface
 public interface WheelEventHandler
 {
   void onWheelEvent( WheelEvent event );